+++ test/fixtures/consumer/lib/dispose_effect_fix_input_inferred.dart
@@ -1,0 +1,14 @@
+import 'package:all_observer/all_observer.dart';
+import 'package:flutter/widgets.dart';
+
+class EffectPageState extends State<StatefulWidget> {
+  late final disposeEffect = effect(() {});
+
+  @override
+  void dispose() {
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) => const SizedBox();
+}